pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "maestro"
version = "0.2.0"
description = "Symphony-compatible coding agent orchestrator powered by Cursor."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "PyYAML>=6.0.1",
  "typer>=0.12.3",
  "python-liquid>=1.12",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
maestro = "maestro.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/maestro"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
